+++ src/wallet/rpcdump.cpp
@@ -16,6 +16,8 @@
 #include "merkleblock.h"
 #include "core_io.h"
 
+#include "rpcwallet.h"
+
 #include <fstream>
 #include <stdint.h>
